+++ lib/src/platform.dart
@@ -45,6 +45,9 @@ abstract class DataSerializerPlatform {
   /// Returns `true` if the platform fully supports 64 bits integers.
   bool get supportsFullInt64;
 
+  /// Returns `true` if the platform fully supports bits shift.
+  bool get supportsFullBitsShift;
+
   /// Returns `true` if [n] is a safe `int`.
   bool isSafeInteger(int n);
 
@@ -115,4 +118,10 @@ abstract class DataSerializerPlatform {
 
   /// Reads a `Int64` from [out] at [offset] in [endian] order.
   int readInt64(Uint8List out, [int offset = 0, Endian endian = Endian.big]);
+
+  /// Performs a right bit shift (`>>`).
+  int shiftRightInt(int n, int shift);
+
+  /// Performs a left bit shift (`<<`).
+  int shiftLeftInt(int n, int shift);
 }