+++ lib/src/platform_generic.dart
@@ -36,6 +36,9 @@ class DataSerializerPlatformGeneric extends DataSerializerPlatform {
   bool get supportsFullInt64 => false;
 
   @override
+  bool get supportsFullBitsShift => false;
+
+  @override
   bool isSafeInteger(int n) {
     return n <= _maxSafeInt && n >= _minSafeInt;
   }
@@ -199,6 +202,30 @@ class DataSerializerPlatformGeneric extends DataSerializerPlatform {
   @override
   int readInt64(Uint8List out, [int offset = 0, Endian endian = Endian.big]) =>
       getInt64(out.asByteData(), offset, endian);
+
+  @override
+  int shiftRightInt(int n, int shift) {
+    if (n >= 0) {
+      return n >> shift;
+    }
+
+    var i = BigInt.from(n);
+    i = i >> shift;
+
+    return i.toInt();
+  }
+
+  @override
+  int shiftLeftInt(int n, int shift) {
+    if (n >= 0) {
+      return n << shift;
+    }
+
+    var i = BigInt.from(n);
+    i = i << shift;
+
+    return i.toInt();
+  }
 }
 
 DataSerializerPlatform createPlatformInstance() =>