+++ lib/src/platform_io.dart
@@ -33,6 +33,9 @@ class DataSerializerPlatformIO extends DataSerializerPlatform {
   int get safeIntBits => 64;
 
   @override
+  bool get supportsFullBitsShift => true;
+
+  @override
   bool get supportsFullInt64 => true;
 
   @override
@@ -132,6 +135,16 @@ class DataSerializerPlatformIO extends DataSerializerPlatform {
     }
     data.setUint32(offset, n0, endian);
     data.setUint32(offset + 4, n1, endian);
+  }
+
+  @override
+  int shiftRightInt(int n, int shift) {
+    return n >> shift;
+  }
+
+  @override
+  int shiftLeftInt(int n, int shift) {
+    return n << shift;
   }
 }